compute/requirements.txt
# Pinned-version requirements for the ECI numerical campaign.
# Use `pip install -r requirements.txt` after `python -m venv ~/.venv/physics`.
# JAX cuda12 is installed separately (see setup-vastai.sh) because the wheel
# selector depends on the host CUDA driver.

# Core numerical
numpy<2
scipy
sympy
mpmath

# Plotting / I/O
matplotlib
seaborn
corner
getdist
h5py
pandas
tqdm
rich

# Testing
pytest
pytest-cov
hypothesis

# MCMC stack
cobaya>=3.5
cosmopower-jax>=0.5
anesthetic
chainconsumer
emcee

# Tensor networks (C3)
tenpy
quimb
opt-einsum

# QFT / GR helpers (C2, C5)
einops
optax       # for JAX-based MCMC if we go that way
jaxopt
